+++ server/config/toml.go
@@ -28,6 +28,30 @@ min-tip = {{ .EVM.MinTip }}
 # GethMetricsAddress defines the addr to bind the geth metrics server to. Default 127.0.0.1:8100.
 geth-metrics-address = "{{ .EVM.GethMetricsAddress }}"
 
+# Mempool configuration for EVM transactions
+[evm.mempool]
+
+# PriceLimit is the minimum gas price to enforce for acceptance into the pool (in wei)
+price-limit = {{ .EVM.Mempool.PriceLimit }}
+
+# PriceBump is the minimum price bump percentage to replace an already existing transaction (nonce)
+price-bump = {{ .EVM.Mempool.PriceBump }}
+
+# AccountSlots is the number of executable transaction slots guaranteed per account
+account-slots = {{ .EVM.Mempool.AccountSlots }}
+
+# GlobalSlots is the maximum number of executable transaction slots for all accounts
+global-slots = {{ .EVM.Mempool.GlobalSlots }}
+
+# AccountQueue is the maximum number of non-executable transaction slots permitted per account
+account-queue = {{ .EVM.Mempool.AccountQueue }}
+
+# GlobalQueue is the maximum number of non-executable transaction slots for all accounts
+global-queue = {{ .EVM.Mempool.GlobalQueue }}
+
+# Lifetime is the maximum amount of time non-executable transaction are queued
+lifetime = "{{ .EVM.Mempool.Lifetime }}"
+
 ###############################################################################
 ###                           JSON RPC Configuration                        ###
 ###############################################################################